+++ examples/plot_dem_hillshade.py
@@ -14,21 +14,26 @@ Learn how to create a hillshade Layer from a DEM using EarthPy
 #       function from EarthPy.
 
 ###############################################################################
-# Create hillshade from DEM
-# -----------------------------
-# Creating a hillshade layer can be a great way to help visualize elevation
-# data, and make your map more presentable. This vignette will show you
-# how to create a hillshade layer using EarthPy, and the different ways to
-# modify the layer to get the hillshade to look exactly how you want it to.
+# Create a Hillshade from a Digital Elevation Model (DEM)
+# -------------------------------------------------------
+# A hillshade is a 3D representation of a surface. Hillshades are generally
+# rendered in greyscale. The darker and lighter colors represent the shadows
+# and highlights that you would visually expect to see in a terrain model.
+# Hillshades are often used as an underlay in a map, to make the data appear
+# more 3-Dimensional and thus visually interesting. This vignette will show
+# you how to create a hillshade from a DEM using EarthPy. It will highlight
+# how to adjust the azimuth and other settings that will impact how the
+# shadows are modeled in the data.
 #
-# To begin using the EarthPy ``hillshade()`` function, import the needed
+# The hillshade function is a part of the spatial module in EarthPy.
+# To begin using the EarthPy hillshade() function, import the needed
 # packages and create an array to be plotted.
 
 ###############################################################################
 # Import Packages
-# ------------------------------
+# ----------------
 #
-# You will need several packages to plot hillshade. You will use RasterIO to
+# You will need several packages to plot hillshade. You will use Rasterio to
 # open up the DEM file you will use to create the hillshade layer. You will
 # primarily be using the EarthPy spatial module in this vignette.
 
@@ -40,92 +45,82 @@ import earthpy.spatial as es
 import earthpy.plot as ep
 import rasterio as rio
 
-# Downloading the data needed for this vignette
-
+# Download the data needed for this vignette
 data = et.data.get_data("colorado-flood")
 
 ####################################################################################
-# Opening up the DEM layer
+# Open up the DEM layer
 # -------------------------
-# To create a hillshade layer, we first need a DEM layer opened up as a numpy array.
-# We can do this with RasterIO. We will read in the elevation data, and take out any
-# data that is bad by setting values below 0 to numpy's not a number value. We then
+# To create a hillshade layer, you first need a DEM layer opened up as a numpy array.
+# You can do this with RasterIO. You will read in the elevation data, and take out any
+# data that is bad by setting values below 0 to numpy's not a number value. You then
 # plot the data with ``ep.plot_bands()`` to show that the data was read in properly.
 
-# Setting the home directory and getting the data for the exercise
-
+# Set the home directory and get the data for the exercise
 os.chdir(os.path.join(et.io.HOME, "earth-analytics"))
 dtm = "data/colorado-flood/spatial/boulder-leehill-rd/pre-flood/lidar/pre_DTM.tif"
 
-# Opening the DEM with RasterIO
-
+# Open the DEM with RasterIO
 with rio.open(dtm) as src:
     elevation = src.read(1)
-    # Fixing bad values
+    # Fix bad values
     elevation[elevation < 0] = np.nan
 
-# Plotting the data
-
+# Plott the data
 ep.plot_bands(elevation, scale=False, cmap="gist_earth")
 plt.show()
 
 ####################################################################################
-# Creating the hillshade layer
+# Create the hillshade layer
 # ----------------------------
 # Once the DEM numpy array is created, simply use ``es.hillshade()`` with the DEM
 # numpy array as an input to create the hillshade layer.
 
-# Creating the hillshade layer
-
+# Create the hillshade layer
 hillshade = es.hillshade(elevation)
 
-# Plotting the hillshade layer
-
+# Plot the hillshade layer
 ep.plot_bands(hillshade, scale=False, cbar=False)
 plt.show()
 
 ####################################################################################
-# Changing the azimuth of the sun
+# Change the azimuth of the sun
 # -------------------------------
 # In hillshade, the angle that the light hits the DEM at will change how the output
 # layer looks. You can adjust this by setting the azimuth angle. Azimuth numbers can
 # range from 0 to 360 degrees, where 0 is due North. The default value for azimuth
 # in ``es.hillshade()`` is 30 degrees.
 
-# Changing the azimuth of the hillshade layer
-
+# Change the azimuth of the hillshade layer
 hillshade_azimuth_210 = es.hillshade(elevation, azimuth=210)
 
-# Plotting the hillshade layer with the modified azimuth
-
+# Plot the hillshade layer with the modified azimuth
 ep.plot_bands(hillshade_azimuth_210, scale=False, cbar=False)
 plt.show()
 
 ####################################################################################
-# Changing the angle altitude of the sun
+# Change the angle altitude of the sun
 # ---------------------------------------
 # Another variable you can adjust for hillshade is what angle altitude the sun is
 # shining from. These values range from 0 to 90, with 90 being the sun shining from
 # directly above the scene. The default value for the angle altitude in
 # ``es.hillshade()`` is 30 degrees.
 
-# Changing the angle altitude of the sun
-
+# Adjust the azimuth value
 hillshade_angle_10 = es.hillshade(elevation, angle_altitude=10)
 
-# Plotting the hillshade layer with the modified angle altitude
-
+# Plot the hillshade layer with the modified angle altitude
 ep.plot_bands(hillshade_angle_10, scale=False, cbar=False)
 plt.show()
 
 ####################################################################################
-# Overlaying a DEM and hillshade layer
+# Overlay a DEM and hillshade layer
 # -------------------------------------
 # One of the benefits of hillshade is to visually enhance DEM layers to add clarity.
 # To do this, the layers have to be plotted together. This can be done very simply
 # with the ``ep.plot_bands()`` function in EarthPy.
 
-# Plotting the DEM and hillshade at the same time
+# Plot the DEM and hillshade at the same time
 # sphinx_gallery_thumbnail_number = 5
 fig, ax = plt.subplots(figsize=(10, 6))
 ep.plot_bands(